refactor: move options stuff to its own file
